requirements.txt
redis==5.0.1
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
